Completed AWGLColor3 tests, fixed 1.0 === 1 bug (HAHAHA)
In js, it is impossible to tell the difference between 1.0 and 1,
unless they are cast to a String. Even then, if not done right,
1.0 loses the '.0' part and the String representation becomes "1".

This means we can't automagically convert to the proper ranges
internally with AWGLColor3, so I've taken that functionality out.

The test for the AWGLColor3 class pointed this out, interestingly
enough, since I decided to test with 1.0. Using 1.2 or anything
off-zero would have worked, so got a tad lucky there.